Move attributes to their own folder
