+++ ConfigLoader/Attributes/ConfigObjectAttribute.cs
@@ -4,7 +4,7 @@
  * You are free to redistribute, share, adapt, etc. as long as the original author (stupid_chris/Christophe Savard) is properly, clearly, *
 \* and explicitly credited, that you do not use this material to a commercial use, and that you distribute it under the same license.     */
 
-namespace ConfigLoader;
+namespace ConfigLoader.Attributes;
 
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct)]
 public class ConfigObjectAttribute : Attribute
