+++ ConfigLoader/ConfigLoader.csproj
@@ -37,8 +37,8 @@
     <DocumentationFile>..\Output\GameData\ConfigLoader\Plugins\ConfigLoader.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="ConfigFieldAttribute.cs" />
-    <Compile Include="ConfigObjectAttribute.cs" />
+    <Compile Include="Attributes\ConfigFieldAttribute.cs" />
+    <Compile Include="Attributes\ConfigObjectAttribute.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\IsExternalInit.cs" />
   </ItemGroup>
